requirements/base.txt
logzero >=1.7.0, <2.0
requests >=2.28.1, <3.0
attrs >=22.1.0, <23.0
cattrs >=22.2.0, <23.0
python-dateutil >=2.8.2, <3.0
click >=8.1.3, <9.0
toml >=0.10.2, <0.11
tabulate >= 0.9.0
jsonschema >= 4.17.0, <5.0
